requirements.txt
streamlit>=1.32.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.21.0
sentence-transformers>=2.6.0
pytesseract>=0.3.10
